[WIP] Basic Setup
Created a User and Server class. The Server listens and can serve
something, in this case a text. The user connects. TDD’d both. Used
commons library to create a command line parser. Jar file is not
working so cob spec tests are failing. Trying to fix it. Can’t get the
apace cli to build with the jar.